Enum ieee80211::ReasonCode

pub enum ReasonCode {

    UnspecifiedReason,
    PreviousAuthenticationNoLongerValid,
    STALeavingIBSSOrESS,
    Inactivity,
    APIsUnableToHandleAllCurrentlyAssociatedSTAs,
    Class2FrameReceivedFromNonauthenticatedSTA,
    Class3FrameReceivedFromNonassociatedSTA,
    STALeavingBSS,
    NotAuthenticated,
    PowerCapabilityElementIsUnacceptable,
    SupportedChannelsElementIsUnacceptable,
    DisassociatedDueToBSSTransitionManagement,
    InvalidInformationElement,
    MessageIntegrityCodeFailure,
    FourWayHandshakeTimeout,
    GroupKeyHandshakeTimeout,
    InformationDifferent,
    InvalidGroupCipher,
    InvalidPairwiseCipher,
    InvalidAKMP,
    UnsupportedRSNInformationElementVersion,
    InvalidRSNInformationElementCapabilities,
    IEEE8021XAuthenticationFailed,
    CipherSuiteRejectedBecauseOfTheSecurityPolicy,
    TDLSDirectLinkTeardownUnreachable,
    TDLSDirectLinkTeardownUnspecifiedReason,
    DisassociatedBecauseSessionTerminatedBySSPRequest,
    DisassociatedBecauseOfLackOfSSPRoamingAgreement,
    RequestedServiceRejected,
    RequestedServiceNotAuthorizedInThisLocation,
    TSDeleted,
    DisassociatedForUnspecifiedQoSRelatedReason,
    QoSAPLacksSufficientBandwidthForThisQoSSTA,
    ExcessiveNumberOfFramesNotAcknowledged,
    STAIsTransmittingOutsideTheLimitsOfItsTXOPs,
    RequestedSTALeavingBSS,
    RequestedSTADoesNotWantToUseTheMechanism,
    RequestedSTAReceivedFrames,
    RequestedSTATimeout,
    PeerSTADoesNotSupportTheRequestedCipherSuite,
    AuthorizedAccessLimitReached,
    DisassociatedDueToExternalServiceRequirements,
    InvalidFTActionFrameCount,
    InvalidPairwiseMasterKeyIdentifier,
    InvalidMDE,
    InvalidFTE,
    SMECancel,
    MeshMaxSupportedMaximumNumberOfPeerMeshSTAs,
    InformationViolatesMeshPolicy,
    MeshSTAReceivedAMeshPeeringCloseMessage,
    MeshSTAHasResentMaxRetriesWithoutConfirm,
    TheConfirmTimerForTheMeshPeeringInstanceTimesOut,
    MeshSTAFailsToUnwrapGTK,
    MeshSTAReceivesInconsistentInformation,
    MeshSTAFailsAuthenticatedMeshPeeringExchange,
    MeshSTADoesNotHaveProxyInformation,
    MeshSTADoesNotHaveForwardingInformation,
    MeshSTALinkNoLongerUsable,
    MacAddressAlreadyExists,
    MeshSTAPerformsChannelSwitchToMeetRegulatoryRequirements,
    MeshSTAPerformsChannelSwitchWithUnspecifiedReason,
    Reserved(u16),

}

Variants

UnspecifiedReason

1 Unspecified reason

PreviousAuthenticationNoLongerValid

2 Previous authentication no longer valid

STALeavingIBSSOrESS

3 Deauthenticated because sending STA is leaving (or has left) IBSS or ESS

Inactivity

4 Disassociated due to inactivity

APIsUnableToHandleAllCurrentlyAssociatedSTAs

5 Disassociated because AP is unable to handle all currently associated STAs

Class2FrameReceivedFromNonauthenticatedSTA

6 Class 2 frame received from nonauthenticated STA

Class3FrameReceivedFromNonassociatedSTA

7 Class 3 frame received from nonassociated STA

STALeavingBSS

8 Disassociated because sending STA is leaving (or has left) BSS

NotAuthenticated

9 STA requesting (re)association is not authenticated with responding STA

PowerCapabilityElementIsUnacceptable

10 Disassociated because the information in the Power Capability element is unacceptable

SupportedChannelsElementIsUnacceptable

11 Disassociated because the information in the Supported Channels element is unacceptable

DisassociatedDueToBSSTransitionManagement

12 Disassociated due to BSS transition management

InvalidInformationElement

13 Invalid information element, i.e., an information element defined in this standard for which the content does not meet the specifications in Clause 7

MessageIntegrityCodeFailure

14 Message integrity code (MIC) failure

FourWayHandshakeTimeout

15 4-Way Handshake timeout

GroupKeyHandshakeTimeout

16 Group Key Handshake timeout

InformationDifferent

17 Information element in 4-Way Handshake different from (Re)Association Request/Probe Response/Beacon frame

InvalidGroupCipher

18 Invalid group cipher

InvalidPairwiseCipher

19 Invalid pairwise cipher

InvalidAKMP

20 Invalid AKMP

UnsupportedRSNInformationElementVersion

21 Unsupported RSN information element version

InvalidRSNInformationElementCapabilities

22 Invalid RSN information element capabilities

IEEE8021XAuthenticationFailed

23 IEEE 802.1X authentication failed

CipherSuiteRejectedBecauseOfTheSecurityPolicy

24 Cipher suite rejected because of the security policy

TDLSDirectLinkTeardownUnreachable

25 TDLS direct-link teardown due to TDLS peer STA unreachable via the TDLS direct link

TDLSDirectLinkTeardownUnspecifiedReason

26 TDLS direct-link teardown for unspecified reason

DisassociatedBecauseSessionTerminatedBySSPRequest

27 Disassociated because session terminated by SSP request

DisassociatedBecauseOfLackOfSSPRoamingAgreement

28 Disassociated because of lack of SSP roaming agreement

RequestedServiceRejected

29 Requested service rejected because of SSP cipher suite or AKM requirement

RequestedServiceNotAuthorizedInThisLocation

30 Requested service not authorized in this location

TSDeleted

31 TS deleted because QoS AP lacks sufficient bandwidth for this QoS STA due to a change in BSS service characteristics or operational mode

DisassociatedForUnspecifiedQoSRelatedReason

32 Disassociated for unspecified, QoS-related reason

QoSAPLacksSufficientBandwidthForThisQoSSTA

33 Disassociated because QoS AP lacks sufficient bandwidth for this QoS STA

ExcessiveNumberOfFramesNotAcknowledged

34 Disassociated because excessive number of frames need to be acknowledged, but are not acknowledged due to AP transmissions and/or poor channel conditions

STAIsTransmittingOutsideTheLimitsOfItsTXOPs

35 Disassociated because STA is transmitting outside the limits of its TXOPs

RequestedSTALeavingBSS

36 Requested from peer STA as the STA is leaving the BSS (or resetting)

RequestedSTADoesNotWantToUseTheMechanism

37 Requested from peer STA as it does not want to use the mechanism

RequestedSTAReceivedFrames

38 Requested from peer STA as the STA received frames using the mechanism for which a setup is required

RequestedSTATimeout

39 Requested from peer STA due to timeout

PeerSTADoesNotSupportTheRequestedCipherSuite

45 Peer STA does not support the requested cipher suite

AuthorizedAccessLimitReached

46 Disassociated because authorized access limit reached

DisassociatedDueToExternalServiceRequirements

47 Disassociated due to external service requirements

InvalidFTActionFrameCount

48 Invalid FT Action frame count

InvalidPairwiseMasterKeyIdentifier

49 Invalid pairwise master key identifier (PMKI)

InvalidMDE

50 Invalid MDE

InvalidFTE

51 Invalid FTE

SMECancel

52 SME cancels the mesh peering instance with the reason other than reaching the maximum number of peer mesh STAs

MeshMaxSupportedMaximumNumberOfPeerMeshSTAs

53 The mesh STA has reached the supported maximum number of peer mesh STAs

InformationViolatesMeshPolicy

54 The received information violates the Mesh Configuration policy configured in the mesh STA profile

MeshSTAReceivedAMeshPeeringCloseMessage

55 The mesh STA has received a Mesh Peering Close message requesting to close the mesh peering

MeshSTAHasResentMaxRetriesWithoutConfirm

56 The mesh STA has re-sent dot11MeshMaxRetries Mesh Peering Open messages, without receiving a Mesh Peering Confirm message

TheConfirmTimerForTheMeshPeeringInstanceTimesOut

57 The confirmTimer for the mesh peering instance times out

MeshSTAFailsToUnwrapGTK

58 The mesh STA fails to unwrap the GTK or the values in the wrapped contents do not match

MeshSTAReceivesInconsistentInformation

59 The mesh STA receives inconsistent information about the mesh parameters between Mesh Peering Management frames

MeshSTAFailsAuthenticatedMeshPeeringExchange

60 The mesh STA fails the authenticated mesh peering exchange because due to failure in selecting either the pairwise ciphersuite or group ciphersuite

MeshSTADoesNotHaveProxyInformation

61 The mesh STA does not have proxy information for this external destination

MeshSTADoesNotHaveForwardingInformation

62 The mesh STA does not have forwarding information for this destination

MeshSTALinkNoLongerUsable

63 The mesh STA determines that the link to the next hop of an active path in its forwarding information is no longer usable

MacAddressAlreadyExists

64 The Deauthentication frame was sent because the MAC address of the STA already exists in the mesh BSS. See 11.3.3 (Additional mechanisms for an AP collocated with a mesh STA)

MeshSTAPerformsChannelSwitchToMeetRegulatoryRequirements

65 The mesh STA performs channel switch to meet regulatory requirements

MeshSTAPerformsChannelSwitchWithUnspecifiedReason

66 The mesh STA performs channel switch with unspecified reason

Reserved(u16)

Implementations

impl ReasonCode

pub fn from_u16(n: u16) -> Self

pub fn into_u16(self) -> u16