Enum ieee80211::StatusCode

pub enum StatusCode {
    Successful,
    UnspecifiedFailure,
    TDLSWakeupScheduleRejectedButAlternativeScheduleProvided,
    TDLSWakeupScheduleRejected,
    SecurityDisabled,
    UnacceptableLifetime,
    NotInSameBSS,
    CannotSupportAllRequestedCapabilities,
    ReassociationDenied,
    AssociationDenied,
    STADoesNotSupportTheSpecifiedAuthenticationAlgorithm,
    ReceivedBadAuthenticationFrame,
    AuthenticationRejectedBecauseOfChallengeFailure,
    AuthenticationRejectedDueToTimeoutWaitingForNextFrameInSequence,
    AssociationDeniedBecauseAPIsUnableToHandleAdditionalAssociatedSTAs,
    AssociationDeniedDataRates,
    AssociationDeniedShortPreamble,
    AssociationDeniedPBCCModulation,
    AssociationDeniedChannelAgility,
    AssociationRequestRejectedSpectrumManagement,
    AssociationRequestRejectedPowerCapability,
    AssociationRequestRejectedSupportedChannels,
    AssociationDeniedDueToRequestingSTANotSupportingTheShortSlotTimeOption,
    AssociationDeniedDueToRequestingSTANotSupportingTheDSSSOFDMOption,
    ReservedAssociationDeniedBecauseTheRequestingSTADoesNotSupportHTFeatures,
    R0KHUnreachable,
    AssociationDeniedPhasedCoexistence,
    AssociationRequestRejectedTemporarily,
    RobustManagementFramePolicyViolation,
    UnspecifiedQoSRelatedFailure,
    AssociationDeniedBecauseInsufficientBandwidth,
    AssociationDeniedExcessiveFrameLoss,
    AssociationWithQoSBSSDenied,
    TheRequestHasBeenDeclined,
    TheRequestHasNotBeenSuccessfulAsOneOrMoreParametersHaveInvalidValues,
    TheAllocationOrTSHasNotBeenCreated,
    InvalidInformationElement,
    InvalidGroupCipher,
    InvalidPairwiseCipher,
    InvalidAKMP,
    UnsupportedRSNInformationElementVersion,
    InvalidRSNInformationElementCapabilities,
    CipherSuiteRejectedBecauseOfSecurityPolicy,
    TheTSPerAllocationHasNotBeenCreated,
    DirectLinkIsNotAllowedInTheBSSByPolicy,
    TheDestinationSTAIsNotPresentWithinThisBSS,
    TheDestinationSTAIsNotAQoSSTA,
    AssociationDeniedBecauseTheListenIntervalIsTooLarge,
    InvalidFTActionFrameCount,
    InvalidPairwiseMasterKeyIdentifier,
    InvalidMDIE,
    InvalidFTIE,
    RequestedTCLASProcessingIsNotSupportedByThePCPOrAP,
    ThePCPOrAPHasInsufficientTCLASProcessingResourcesToSatisfyTheRequest,
    TheTSHasNotBeenCreatedBecauseTheRequestCannotBeHonored,
    GASAdvertisementProtocolNotSupported,
    NoOutstandingGASRequest,
    GASResponseNotReceivedFromTheAdvertisementServer,
    STATimedOutWaitingForGASQueryResponse,
    GASResponseIsLargerThanQueryResponseLengthLimit,
    RequestRefusedBecauseHomeNetworkDoesNotSupportRequest,
    AdvertisementServerInTheNetworkIsNotCurrentlyReachable,
    RequestRefusedDueToPermissionsReceivedViaSSPNInterface,
    RequestRefusedBecausePCPOrAPDoesNotSupportUnauthenticatedAccess,
    InvalidContentsOfRSNIE,
    UAPSDCoexistenceIsNotSupported,
    RequestedUAPSDCoexistenceModeIsNotSupported,
    RequestedIntervalDurationValueCannotBeSupportedWithUAPSDCoexistence,
    AuthenticationIsRejectedBecauseAnAntiCloggingTokenIsRequired,
    AuthenticationIsRejectedBecauseTheOfferedFiniteCyclicGroupIsNotSupported,
    TheTBTTAdjustmentRequestHasNotBeenSuccessful,
    TransmissionFailure,
    RequestedTCLASNotSupported,
    TCLASResourcesExhausted,
    RejectedWithSuggestedBSSTransition,
    RejectWithRecommendedSchedule,
    RejectBecauseNoWakeupScheduleSpecified,
    SuccessTheDestinationSTAIsInPowerSaveMode,
    FSTPendingInProcessOfAdmittingFSTSession,
    PerformingFSTNow,
    FSTPendingGapInBlockAckWindow,
    RejectBecauseOfUPIDSetting,
    AssociationRefusedForSomeExternalReason,
    AssociationRefusedBecauseOfMemoryLimitsAtTheAP,
    AssociationRefusedBecauseEmergencyServicesAreNotSupportedAtTheAP,
    GASQueryResponseNotYetReceived,
    RejectDSEProcedures,
    TheAssociationHasBeenDenied,
    TheRequestFailedDueToAReservationConflict,
    TheRequestFailedDueToExceededMAFLimit,
    TheRequestFailedDueToExceededMCCATrackLimit,
    AssociationDeniedBecauseSpectrumManagement,
    AssociationDeniedBecauseTheRequestingSTADoesNotSupportVHTFeatures,
    EnablementDenied,
    EnablementDeniedDueToRestrictionFromAnAuthorizedGDB,
    AuthorizationDeenabled,
    AuthenticationRejectedDueToFILSAuthenticationFailure,
    AuthenticationRejectedDueToUnknownAuthenticationServer,
    Reserved(u16),
}

Variants

Successful

0 Successful

UnspecifiedFailure

1 Unspecified failure

TDLSWakeupScheduleRejectedButAlternativeScheduleProvided

2 TDLS wakeup schedule rejected but alternative schedule provided

TDLSWakeupScheduleRejected

3 TDLS wakeup schedule rejected

SecurityDisabled

4 Reserved 5 Security disabled

UnacceptableLifetime

6 Unacceptable lifetime

NotInSameBSS

7 Not in same BSS

CannotSupportAllRequestedCapabilities

8 Reserved 9 Reserved 10 Cannot support all requested capabilities in the Capability Information field

ReassociationDenied

11 Reassociation denied due to inability to confirm that association exists

AssociationDenied

12 Association denied due to reason outside the scope of this standard

STADoesNotSupportTheSpecifiedAuthenticationAlgorithm

13 Responding STA does not support the specified authentication algorithm

ReceivedBadAuthenticationFrame

14 Received an Authentication frame with authentication transaction sequence number out of expected sequence

AuthenticationRejectedBecauseOfChallengeFailure

15 Authentication rejected because of challenge failure

AuthenticationRejectedDueToTimeoutWaitingForNextFrameInSequence

16 Authentication rejected due to timeout waiting for next frame in sequence

AssociationDeniedBecauseAPIsUnableToHandleAdditionalAssociatedSTAs

17 Association denied because AP is unable to handle additional associated STAs

AssociationDeniedDataRates

18 Association denied due to requesting STA not supporting all of the data rates in the BSSBasicRateSet parameter

AssociationDeniedShortPreamble

19 Association denied due to requesting STA not supporting the short preamble option

AssociationDeniedPBCCModulation

20 Association denied due to requesting STA not supporting the PBCC modulation option

AssociationDeniedChannelAgility

21 Association denied due to requesting STA not supporting the Channel Agility option

AssociationRequestRejectedSpectrumManagement

22 Association request rejected because Spectrum Management capability is required

AssociationRequestRejectedPowerCapability

23 Association request rejected because the information in the Power Capability element is unacceptable

AssociationRequestRejectedSupportedChannels

24 Association request rejected because the information in the Supported Channels element is unacceptable

AssociationDeniedDueToRequestingSTANotSupportingTheShortSlotTimeOption

25 Association denied due to requesting STA not supporting the Short Slot Time option

AssociationDeniedDueToRequestingSTANotSupportingTheDSSSOFDMOption

26 Association denied due to requesting STA not supporting the DSSS-OFDM option

ReservedAssociationDeniedBecauseTheRequestingSTADoesNotSupportHTFeatures

27 Reserved Association denied because the requesting STA does not support HT features

R0KHUnreachable

28 R0KH unreachable

AssociationDeniedPhasedCoexistence

29 Association denied because the requesting STA does not support the phased coexistence operation (PCO) transition time required by the AP

AssociationRequestRejectedTemporarily

30 Association request rejected temporarily; try again later

RobustManagementFramePolicyViolation

31 Robust Management frame policy violation

UnspecifiedQoSRelatedFailure

32 Unspecified, QoS-related failure

AssociationDeniedBecauseInsufficientBandwidth

33 Association denied because QoS AP or PCP has insufficient bandwidth to handle another QoS STA

AssociationDeniedExcessiveFrameLoss

34 Association denied due to excessive frame loss rates and/or poor conditions on current operating channel

AssociationWithQoSBSSDenied

35 Association (with QoS BSS) denied because the requesting STA does not support the QoS facility

TheRequestHasBeenDeclined

36 Reserved 37 The request has been declined

TheRequestHasNotBeenSuccessfulAsOneOrMoreParametersHaveInvalidValues

38 The request has not been successful as one or more parameters have invalid values

TheAllocationOrTSHasNotBeenCreated

39 The allocation or TS has not been created because the request cannot be honored; however, a suggested TSPEC/DMG TSPEC is provided so that the initiating STA may attempt to set another allocation or TS with the suggested changes to the TSPEC/DMG TSPEC

InvalidInformationElement

40 Invalid information element, i.e., an information element defined in this standard for which the content does not meet the specifications in Clause 7

InvalidGroupCipher

41 Invalid group cipher

InvalidPairwiseCipher

42 Invalid pairwise cipher

InvalidAKMP

43 Invalid AKMP

UnsupportedRSNInformationElementVersion

44 Unsupported RSN information element version

InvalidRSNInformationElementCapabilities

45 Invalid RSN information element capabilities

CipherSuiteRejectedBecauseOfSecurityPolicy

46 Cipher suite rejected because of security policy

TheTSPerAllocationHasNotBeenCreated

47 The TS per allocation has not been created; however, the PCP/HC may be capable of creating a TS or allocation, in response to a request, after the time indicated in the TS Delay element

DirectLinkIsNotAllowedInTheBSSByPolicy

48 Direct link is not allowed in the BSS by policy

TheDestinationSTAIsNotPresentWithinThisBSS

49 The Destination STA is not present within this BSS

TheDestinationSTAIsNotAQoSSTA

50 The Destination STA is not a QoS STA

AssociationDeniedBecauseTheListenIntervalIsTooLarge

51 Association denied because the ListenInterval is too large

InvalidFTActionFrameCount

52 Invalid FT Action frame count

InvalidPairwiseMasterKeyIdentifier

53 Invalid pairwise master key identifier (PMKID)

InvalidMDIE

54 Invalid MDIE

InvalidFTIE

55 Invalid FTIE

RequestedTCLASProcessingIsNotSupportedByThePCPOrAP

56 Requested TCLAS processing is not supported by the PCP/AP

ThePCPOrAPHasInsufficientTCLASProcessingResourcesToSatisfyTheRequest

57 The PCP/AP has insufficient TCLAS processing resources to satisfy the request

TheTSHasNotBeenCreatedBecauseTheRequestCannotBeHonored

58 The TS has not been created because the request cannot be honored; however, the PCP/HC suggests the STA to transition to other BSSs to setup the TS

GASAdvertisementProtocolNotSupported

59 GAS Advertisement Protocol not supported

NoOutstandingGASRequest

60 No outstanding GAS request

GASResponseNotReceivedFromTheAdvertisementServer

61 GAS Response not received from the Advertisement Server

STATimedOutWaitingForGASQueryResponse

62 STA timed out waiting for GAS Query Response

GASResponseIsLargerThanQueryResponseLengthLimit

63 GAS Response is larger than query response length limit

RequestRefusedBecauseHomeNetworkDoesNotSupportRequest

64 Request refused because home network does not support request

AdvertisementServerInTheNetworkIsNotCurrentlyReachable

65 Advertisement Server in the network is not currently reachable

RequestRefusedDueToPermissionsReceivedViaSSPNInterface

66 Reserved 67 Request refused due to permissions received via SSPN interface

RequestRefusedBecausePCPOrAPDoesNotSupportUnauthenticatedAccess

68 Request refused because PCP/AP does not support unauthenticated access

InvalidContentsOfRSNIE

69 Reserved 70 Reserved 71 Reserved 72 Invalid contents of RSNIE

UAPSDCoexistenceIsNotSupported

73 U-APSD Coexistence is not supported

RequestedUAPSDCoexistenceModeIsNotSupported

74 Requested U-APSD Coexistence mode is not supported

RequestedIntervalDurationValueCannotBeSupportedWithUAPSDCoexistence

75 Requested Interval/Duration value cannot be supported with U-APSD Coexistence

AuthenticationIsRejectedBecauseAnAntiCloggingTokenIsRequired

76 Authentication is rejected because an Anti-Clogging Token is required

AuthenticationIsRejectedBecauseTheOfferedFiniteCyclicGroupIsNotSupported

77 Authentication is rejected because the offered finite cyclic group is not supported

TheTBTTAdjustmentRequestHasNotBeenSuccessful

78 The TBTT adjustment request has not been successful because the STA could not find an alternative TBTT

TransmissionFailure

79 Transmission failure

RequestedTCLASNotSupported

80 Requested TCLAS Not Supported

TCLASResourcesExhausted

81 TCLAS Resources Exhausted

RejectedWithSuggestedBSSTransition

82 Rejected with Suggested BSS Transition

RejectWithRecommendedSchedule

83 Reject with recommended schedule

RejectBecauseNoWakeupScheduleSpecified

84 Reject because no wakeup schedule specified

SuccessTheDestinationSTAIsInPowerSaveMode

85 Success, the destination STA is in power save mode

FSTPendingInProcessOfAdmittingFSTSession

86 FST pending, in process of admitting FST session

PerformingFSTNow

87 performing FST now

FSTPendingGapInBlockAckWindow

88 FST pending, gap(s) in Block Ack window

RejectBecauseOfUPIDSetting

89 Reject because of U-PID setting

AssociationRefusedForSomeExternalReason

92 (Re)association refused for some external reason

AssociationRefusedBecauseOfMemoryLimitsAtTheAP

93 (Re)association refused because of memory limits at the AP

AssociationRefusedBecauseEmergencyServicesAreNotSupportedAtTheAP

94 (Re)association refused because emergency services are not supported at the AP

GASQueryResponseNotYetReceived

95 GAS query response not yet received

RejectDSEProcedures

96 Reject since the request is for transition to a frequency band subject to DSE procedures and FST initiator is a dependent STA

TheAssociationHasBeenDenied

97 Reserved 98 Reserved 99 The association has been denied; however, one or more Multi-band elements are included that can be used by the receiving STA to join the BSS

TheRequestFailedDueToAReservationConflict

100 The request failed due to a reservation conflict

TheRequestFailedDueToExceededMAFLimit

101 The request failed due to exceeded MAF limit

TheRequestFailedDueToExceededMCCATrackLimit

102 The request failed due to exceeded MCCA track limit

AssociationDeniedBecauseSpectrumManagement

103 Association denied because the information in the Spectrum Management field is unacceptable

AssociationDeniedBecauseTheRequestingSTADoesNotSupportVHTFeatures

104 Association denied because the requesting STA does not support VHT features

EnablementDenied

105 Enablement denied

EnablementDeniedDueToRestrictionFromAnAuthorizedGDB

106 Enablement denied due to restriction from an authorized GDB

AuthorizationDeenabled

107 Authorization deenabled

AuthenticationRejectedDueToFILSAuthenticationFailure

112 Authentication rejected due to FILS authentication failure

AuthenticationRejectedDueToUnknownAuthenticationServer

113 Authentication rejected due to unknown Authentication Server

Reserved(u16)

Implementations

impl StatusCode

#[must_use]pub fn from_u16(n: u16) -> Self

#[must_use]pub fn into_u16(self) -> u16